feat: add ability to get commit scopes from a config file inside a repo
This commit transplants some of the logic from older implementation that
allows checking a repository for a config file and getting the commit
types from that config file.

The logic handles absence of a repo or a config file inside the repo by
returning the list of default commit types.

Config parser is implemented in a slightly custom way that allows
turning key-value pairs in toml into name-description pairs that
comprise a commit type.