feat: expose ability to get commit scopes to the cli
These changes expose feature from d635130 to the command line.

The cli now looks for a repository in the current directory and tries to
get the custom commit types for it.

During hands-on testing, I added a number of debug statements so that
the program explains the decision tree to the user if the debug flag is
passed.

The change is covered by an integration test. Setting up the test
required factoring out fake config generation into the lib (similar to
0ff30d9).