feat: add scopes output from history
Transplanted the mechanism to extract scopes from the git commit
history.

The implementation of message parsing is suboptimal due to the need of
a lookup of a git message in the tree. This can be probably implemented
using a different mechanism later.

The scope cli argument now produces a merged set of the scopes from
history and from config. Since config ones have description attached to
them, when merging -- duplicates from history are not added.

There is a small refactor that touches lib.rs to make the fixture
generation more, well, generic.