test: add fixture function that creates a test repo with file control
The new `setup_repo_with_commits_and_files` is basically old
`setup_repo_with_commits` that now also accepts a vector of files.

The commit messages are to be zipped with files and the commit message
will be added as a text to the corresponding file.

This is done for the more fine-grained testing that is done for the
refactored scope X commit extraction.

`setup_repo_with_commits` is an alias of the new function that just
dumps everything into a single file. Old tests have never cared for the
files anyway and their code remains the same.