feat: suggest a commit scope that is similar to staged files
This set of changes makes the program examine currently staged set of
files and compare it against past changes that were tied to scopes. If a
matching scope is found, it is rendered first in the list of
suggestions.

The implementation is a bit naive -- it just does strict string
comparison at the time.

Non-functional:

* Add an integration test to check that a commit scope matching the
  currently staged files is pushed to the front
* I added a default method for PrintableEntity -- this allows generating
  a new instance of this struct without specifying description