build: add cargo tests to the flake check
I want `nix flake check` to be the main entry pointo into the CI checks.

Cranelib has a pattern where a package can be added to the list of
checks. `nix flake check` then will run the cargo test and report any
errors.

I moved `default` package definition up to let-binding so that it can be
reused in `checks`