fix(scope): detect partially staged files properly
Prior to this change, files that were not completely staged did not pass
the filter defined in `get_staged_files`. Their status was
Status(INDEX_MODIFIED | WT_MODIFIED) and the check was only for
INDEX_MODIFIED.

I could not just add WT_MODIFIED to the list because that would catch
unstaged files too and did not work anyway since ("a | b" != "a") and
("a|b" != "b").

The new approach is to use a list of allowed statuses and search through
it. While order does not matter, an array is very comparable to a
HashSet for small lookups.