refactor: add repo validation logic to the beginning of the run
The program uses a path to pass a repository around. Which results in a
lot of calls to Repository::discover in the codebase.

I want to clean it up and pass the actual non-bare repository around.

This commit starts this process by inserting the repo check at the
beginning of the program.